Move BlockSubsidyProvider to separate file
