+++ divi/src/SuperblockHelpers.h
@@ -14,21 +14,6 @@ bool IsValidTreasuryBlockHeight(int nBlockHeight);
 
 CBlockRewards GetBlockSubsidity(int nHeight);
 
-class BlockSubsidyProvider: public I_BlockSubsidyProvider
-{
-private:
-    const CChainParams& chainParameters_;
-    I_SuperblockHeightValidator& heightValidator_;
-
-    void updateTreasuryReward(int nHeight, CBlockRewards& rewards, bool isTreasuryBlock) const;
-    void updateLotteryReward(int nHeight, CBlockRewards& rewards,bool isLotteryBlock) const;
-public:
-    BlockSubsidyProvider(
-        const CChainParams& chainParameters,
-        I_SuperblockHeightValidator& heightValidator);
-    virtual CBlockRewards GetBlockSubsidity(int nHeight) const;
-    virtual CAmount GetFullBlockValue(int nHeight) const;
-};
 
 class SuperblockSubsidyProvider
 {
